Increase verbosity of RAP plugin compilation and reproduction.
The original version should have been fine as links to the source code
were included in the original blog post and referenced from the
repository.

After continuous complaints by PaXTeam about GPL violations, we now
also include the patch itself, a link to the patch, and a note on how
to compile the plugin yourself.

Happy GPL and reproduction :)